boot.vc(hierfstat)	R Documentation

Bootstrap confidence intervals for variance components

Description

Provides a bootstrap confidence interval (over loci) for sums of the different variance components (equivalent to gene diversity estimates at the different levels), and the derived F-statistics, as suggested by Weir and Cockerham (1984). Will not run with less than 5 loci. Raymond and Rousset (199X) points out shortcomings of this method.

Usage

boot.vc(levels=levels,loci=loci,diploid=TRUE,nboot=1000,quant=c(0.025,0.5,0.975))

Arguments

levels 	 a data frame containing the different levels (factors) from the outermost (e.g. region) to the innermost before the individual

loci 	 a data frame containing the different loci

diploid 	 Specify whether the data are coming from diploid or haploid organisms (diploid is the default)

nboot 	 Specify the number of bootstrap to carry out. Default is 1000

quant 	 Specify which quantile to produce. Default is c(0.025,0.5,0.975) giving the percentile 95% CI and the median

Value

boot 	 a data frame with the bootstrapped variance components. Could be used for obtaining bootstrap ci of statistics not listed here.

res 	 a data frame with the bootstrap derived statistics. H stands for gene diversity, F for F-statistics

ci 	 Confidence interval for each statistic.

References

Raymond M and Rousset F, 1995. An exact test for population differentiation. Evolution. 49:1280-1283

Weir, B.S. (1996) Genetic Data Analysis II. Sinauer Associates.

Weir BS and Cockerham CC, 1984. Estimating F-statistics for the analysis of population structure. Evolution 38:1358-1370.

See Also

varcomp.glob.

Examples

#load data set
data(gtrunchier)
boot.vc(gtrunchier[,c(1:2)],gtrunchier[,-c(1:2)],nboot=100)

[Package hierfstat version 0.04-1 Index]