genot2al(hierfstat)	R Documentation

Separates diploid genotypes in its constituant alleles

Description

Separates the input vector of diploid genotypes in two vectors each containing one allele, and returns a vector of length 2*length(y) with the second part being the second allele

Usage

genot2al(y)

Arguments

y 	 the diploid genotypes at one locus

Value

returns a vector of length 2*length(y), with the second half of the vector containing the second alleles

Author(s)

Jerome Goudet, DEE, UNIL, CH-1015 Lausanne Switzerland

jerome.goudet@unil.ch

References

Goudet J. (2004). A library for R to compute and test variance components and F-statistics. In Prep

See Also

varcomp.

Examples

data(gtrunchier)
genot2al(gtrunchier[,4])

[Package Contents]