gtrunchier(hierfstat)	R Documentation

Genotypes at 6 microsatellite loci of Galba truncatula from different patches in Western Switzerland

Description

Data set consisting of the microsatellite genotypes of 370 Galba truncatula, a tiny freshwater snail, collecting from different localities and several patches within localities in Western Switzerland.

Usage

data(gtrunchier)

Format

Locality

Identifier of the locality of origin

Patch

Identifier of the patch of origin

L21.V

Genotype at locus L21.V. For instance the first individual carries allele 2 and 2 at this locus
gtrunchier\$L21.V[1]

L37.J

Genotype at locus L37.J

L20.B

Genotype at locus L20.B

L29.V

Genotype at locus L29.V

L36.B

Genotype at locus L36.B

L16.J

Genotype at locus L16.J

References

Trouve S., L. Degen et al. (2000) Microsatellites in the hermaphroditic snail, Lymnaea truncatula, intermediate host of the liver fluke, Fasciola hepatica.Molecular Ecology 9: 1662-1664.

Trouve S., Degen L. and Goudet J. (2005) Ecological components and evolution of selfing in the freshwater snail Galba truncatula. Journal of Evolutionary Biology. 18, 358-370

[Package hierfstat version 0.03-1 Index]