samp.between.within(hierfstat)	R Documentation

Shuffles a sequence

Description

Used to generate a permutation of a sequence 1:length(inner.lev). blocks of observations defined by inner.lev are permutted within blocks defined by outer.lev

Usage

samp.between.within(inner.lev, outer.lev)

Arguments

inner.lev 	 a vector containing the groups to be permuted.

outer.lev 	 a vector containing teh blocks within which observations are to be kept.

Value

a vector 1:length(lev) (with blocks defined by data) randomly permuted. Usually, one passes the result to reorder observations in a data set in order to carry out permutation-based tests

See Also

test.between.within.

[Package Contents]