test.between(hierfstat)	R Documentation

Tests the significance of the effect of test.lev on genetic differentiation

Description

Usage

test.between(data, test.lev, rand.unit, nperm, ...)

Arguments

data 	 a data frame containing the genotypes for the different loci

test.lev 	 A vector containing the units from which to construct the contingency tables

rand.unit 	 A vector containing the assignment of each observation to the units to be permutted

nperm 	 The number of permutations to carry out for the test

... 	 Mainly here to allow passing diploid=FALSE if necessary

Value

g.star 	 A vector containing all the generated g-statistics, the last one beeing the observed

p.val 	 The p-value associated with the test

Author(s)

Jerome Goudet jerome.goudet@unil.ch

Examples

data(gtrunchier)
attach(gtrunchier)
#test whether the locality level has a significant effect on genetic structuring
test.between(gtrunchier[,-c(1,2)],test.lev=Locality,rand.unit=Patch)

[Package Contents]