varcomp(hierfstat)	R Documentation

Estimates variance components for each allele of a locus

Description

Estimates variance components for each allele for a (fully) hierarchical random design defined by all but the last column of the data frame data, the last column containing the genetic data to analyse. Columns for the hierarchical design should be given from the outermost to the innermost before the individual (e.g. continent, region, population, patch,...)

Usage

varcomp(data,diploid=TRUE)

Arguments

data 	 a data frame that contains the different factors from the outermost (e.g. region) to the innermost before the individual. the last column of the data frame 'data' contains the locus to analyse, which can be multiallelic. Missing data are allowed.

diploid 	 a boolean stating whether the data come from diploid (TRUE=default) or haploid (FALSE) organisms

Details

The format for genotypes is simply the code for the 2 alleles put one behind the other, without space in between. For instance if allele 1 at the locus has code 23 and allele 2 39, the genotype format is 2339.

Value

df 	 the degrees of freedom for each level

k 	 the k matrix, the coefficients associated with the variance components

res 	 the variance components for each allele

overall 	 the variance components summed over alleles

F 	 a matrix of hierarchical F-statistics type-coefficients with the first line corresponding to F_{(n-1)/n},F_{(n-2)/n}...F_{i/n} and the diagonal corresponding to F_{(n-1)/n}, F_{(n-2)/(n-1)},F_{i/2}

Author(s)

Jerome Goudet, DEE, UNIL, CH-1015 Lausanne Switzerland

jerome.goudet@unil.ch

http://www.unil.ch/popgen/people/jerome.htm

References

Goudet J. (2005). Hierfstat, a package for R to compute and test variance components and F-statistics. Molecular Ecology Notes. 5:184-186

Weir, B.S. (1996) Genetic Data Analysis II. Sinauer Associates.

Yang, R.C. (1998). Estimating hierarchical F-statistics. Evolution 52(4):950-956

See Also

varcomp.glob.

Examples

#load data set
data(gtrunchier)
attach(gtrunchier)
#
varcomp(data.frame(Locality,Patch,L21.V))

[Package hierfstat version 0.03-2 Index]