varcomp.glob(hierfstat)	R Documentation

Estimate variance components and hierarchical F-statistics over all loci

Description

Return multilocus estimators of variance components and F-statistics

Usage

varcomp.glob(levels=levels,loci=loci,diploid=TRUE)

Arguments

levels 	 a data frame containing the different levels (factors) from the outermost (e.g. region) to the innermost before the individual

loci 	 a data frame containing the different loci

diploid 	 Specify whether the data are coming from diploid or haploid organisms (diploid is the default)

Value

loc 	 The variance components for each locus

overall 	 The variance components summed over all loci

F 	 a matrix of hierarchical F-statistics type-coefficients with the first line corresponding to F_{(n-1)/n},F_{(n-2)/n}...F_{i/n} and the diagonal corresponding to F_{(n-1)/n}, F_{(n-2)/(n-1)},F_{i/2}

Author(s)

Jerome Goudet DEE, UNIL, CH-1015 Lausanne Switzerland

jerome.goudet@unil.ch

References

Weir, B.S. (1996) Genetic Data Analysis II. Sinauer Associates.

Yang, R.C. (1998). Estimating hierarchical F-statistics. Evolution 52(4):950-956

Goudet J. (2005). Hierfstat, a package for R to compute and test variance components and F-statistics. Molecular Ecology Notes. 5:184-186

See Also

varcomp.

Examples

#load data set
data(gtrunchier)
attach(gtrunchier)
varcomp.glob(data.frame(Locality,Patch),gtrunchier[,-c(1,2)])

[Package hierfstat version 0.03-1 Index]